+++ src/Models/CloudEmailConfig.php
@@ -2,7 +2,7 @@
 
 namespace DreamFactory\Core\Models;
 
-class CloudEmailConfig extends BaseServiceConfigModel
+class CloudEmailConfig extends BaseEmailServiceConfigModel
 {
     protected $table = 'cloud_email_config';
 
@@ -14,53 +14,4 @@ class CloudEmailConfig extends BaseServiceConfigModel
     ];
 
     protected $encrypted = ['key'];
-
-    protected $appends = ['parameters'];
-
-    protected $parameters = [];
-
-    public static function boot()
-    {
-        parent::boot();
-
-        static::created(
-            function (CloudEmailConfig $emailConfig){
-                if (!empty($emailConfig->parameters)) {
-                    $params = [];
-                    foreach ($emailConfig->parameters as $param) {
-                        $params[] = new EmailServiceParameterConfig($param);
-                    }
-                    $emailConfig->parameter()->saveMany($params);
-                }
-
-                return true;
-            }
-        );
-    }
-
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\HasMany
-     */
-    public function parameter()
-    {
-        return $this->hasMany(EmailServiceParameterConfig::class, 'service_id');
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getParametersAttribute()
-    {
-        $this->parameters = $this->parameter()->get()->toArray();
-
-        return $this->parameters;
-    }
-
-    /**
-     * @param array $val
-     */
-    public function setParametersAttribute(Array $val)
-    {
-        $this->parameters = $val;
-    }
 }