+++ src/Models/SmtpConfig.php
@@ -4,7 +4,7 @@ namespace DreamFactory\Core\Models;
 
 use DreamFactory\Core\Exceptions\BadRequestException;
 
-class SmtpConfig extends BaseServiceConfigModel
+class SmtpConfig extends BaseEmailServiceConfigModel
 {
     protected $table = 'smtp_config';
 
@@ -19,29 +19,6 @@ class SmtpConfig extends BaseServiceConfigModel
     ];
 
     protected $encrypted = ['username', 'password'];
-
-    protected $appends = ['parameters'];
-
-    protected $parameters = [];
-
-    public static function boot()
-    {
-        parent::boot();
-
-        static::created(
-            function (SmtpConfig $emailConfig){
-                if (!empty($emailConfig->parameters)) {
-                    $params = [];
-                    foreach ($emailConfig->parameters as $param) {
-                        $params[] = new EmailServiceParameterConfig($param);
-                    }
-                    $emailConfig->parameter()->saveMany($params);
-                }
-
-                return true;
-            }
-        );
-    }
 
     /**
      * {@inheritdoc}
@@ -60,31 +37,5 @@ class SmtpConfig extends BaseServiceConfigModel
         }
 
         return true;
-    }
-
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\HasMany
-     */
-    public function parameter()
-    {
-        return $this->hasMany(EmailServiceParameterConfig::class, 'service_id');
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getParametersAttribute()
-    {
-        $this->parameters = $this->parameter()->get()->toArray();
-
-        return $this->parameters;
-    }
-
-    /**
-     * @param array $val
-     */
-    public function setParametersAttribute(Array $val)
-    {
-        $this->parameters = $val;
     }
 }