+++ src/Models/UserLookup.php
@@ -10,7 +10,7 @@ namespace DreamFactory\Core\Models;
  * @property string  $name
  * @property string  $value
  * @property string  $description
- * @property boolean $is_private
+ * @property boolean $private
  * @property string  $created_date
  * @property string  $last_modified_date
  * @method static \Illuminate\Database\Query\Builder|RoleLookup whereId($value)
@@ -28,5 +28,5 @@ class UserLookup extends BaseSystemLookup
 
     protected $fillable = ['user_id', 'name', 'value', 'private', 'description'];
 
-    protected $casts = ['id' => 'integer', 'user_id' => 'integer', 'is_private' => 'boolean'];
+    protected $casts = ['id' => 'integer', 'user_id' => 'integer', 'private' => 'boolean'];
 }