+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [0.1.12] - 2016-01-05
+### Fixed
+- Fix table name case issue with cache lookup for SQL DB services.
+- Fix password setting issue on non-admin user via system/user resource.
+- Fixed additional Email Service parameters from not showing on the admin UI.
+
 ## [0.1.11] - 2015-12-30
 ### Fixed
 - Hotfix for API Doc caching issue.
@@ -95,7 +101,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2015-10-24
 First official release working with the new [dreamfactory](https://github.com/dreamfactorysoftware/dreamfactory) project.
 
-[Unreleased]: https://github.com/dreamfactorysoftware/df-core/compare/0.1.11...HEAD
+[Unreleased]: https://github.com/dreamfactorysoftware/df-core/compare/0.1.12...HEAD
+[0.1.12]: https://github.com/dreamfactorysoftware/df-core/compare/0.1.11...0.1.12
 [0.1.11]: https://github.com/dreamfactorysoftware/df-core/compare/0.1.10...0.1.11
 [0.1.10]: https://github.com/dreamfactorysoftware/df-core/compare/0.1.9...0.1.10
 [0.1.9]: https://github.com/dreamfactorysoftware/df-core/compare/0.1.8...0.1.9